requirements.txt
fastapi==0.115.12
imbalanced_learn==0.13.0
imblearn==0.0
joblib==1.4.2
kaggle==1.7.4.2
matplotlib==3.10.1
mlflow==2.19.0
mlflow_skinny==2.19.0
numpy==2.2.5
omegaconf==2.3.0
pandas==2.2.3
pydantic==2.11.3
python-dotenv==1.1.0
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.1.0
tabulate==0.9.0
xgboost==3.0.0
-e .
